+++ spec/features/events/deleting_events_spec.rb
